requirements.txt
fastapi
uvicorn
google-cloud-storage
requests
python-multipart
python-dotenv
